Fix redudant u32 cast in roundUpPow2

+++ include/math/seadMathCalcCommon.hpp
@@ -477,16 +477,16 @@ inline T MathCalcCommon<T>::roundUp(T x, u32 multNumber)
 template <>
 inline s32 MathCalcCommon<u32>::roundUpPow2(u32 val, u32 base)
 {
-    SEAD_ASSERT_MSG((u32(base - 1) & u32(base)) == 0, "illegal param[val:%d, base:%d]", val, base);
-    return (val + base - 1) & (u32)-base;
+    SEAD_ASSERT_MSG(((base - 1) & base) == 0, "illegal param[val:%d, base:%d]", val, base);
+    return (val + base - 1) & -base;
 }
 
 template <>
 inline s32 MathCalcCommon<s32>::roundUpPow2(s32 val, u32 base)
 {
-    SEAD_ASSERT_MSG(val >= 0 && (u32(base - 1) & u32(base)) == 0, "illegal param[val:%d, base:%d]",
+    SEAD_ASSERT_MSG(val >= 0 && ((base - 1) & base) == 0, "illegal param[val:%d, base:%d]",
                     val, base);
-    return (val + base - 1) & (u32)-base;
+    return (val + base - 1) & -base;
 }
 
 template <typename T>
